sql: Add statement_timeout session setting
From the Postgres docs, this setting does the following:
>Abort any statement that takes more than the specified number of
>milliseconds, starting from the time the command arrives at the server
>from the client. If log_min_error_statement is set to ERROR or lower,
>the statement that timed out will also be logged. A value of zero (the
>default) turns this off.

This setting works slightly differently here, because rather than
starting the timer when the statement arrives at the server, the timer
starts right after the query is added to ActiveQueries. This is to avoid
races between cancellation and execution.

Fixes #20789.

Release note (sql change): Add support for query timeouts with the
PG-compatible `statement_timeout` session setting.